major update: motivation parameter setup
in old version, motivation parameter is equal to reward.
in this version, motivation parameter is scaled by a scaling parameter. By default, the scaling parameter is 1, so m = received reward. if this scaling parameter is set to 10, this will set a very high motivation parameter, even though the ultimate reward is still 2.